Reworked fault get
Refactored some int to long

+++ client/impl/src/main/java/it/xeniaprogetti/cisco/ucs/plugin/client/impl/model/response/AaaRefreshResponse.java
@@ -26,7 +26,7 @@ public class AaaRefreshResponse extends UcsXmlApiResponse {
     @JacksonXmlProperty(isAttribute = true)
     public String outPasswdExpiryStatus;
     @JacksonXmlProperty(isAttribute = true)
-    public int outPasswdExpiryDuration;
+    public long outPasswdExpiryDuration;
 
     @Override
     public String toString() {
